refactor: share jev types across processes, store money as centavos.

- jev domain and ipc types moved to electron/shared, compiled by both sides
- read dates are YYYY-MM-DD strings, write dates stay Date
- debit and credit stored as integer centavos instead of floats
- added protected route for /main
- fixed list paging, failed fetch handling and journal type switching
- contexts split into provider and hook files, eslint clean

Existing databases keep their REAL columns and must be deleted.

+++ electron/db/database.ts
@@ -81,8 +81,9 @@ const createTables = (): void => {
       sort_order        INTEGER NOT NULL,
       account_code      TEXT    NOT NULL,
       account_name      TEXT    NOT NULL,
-      debit             REAL,
-      credit            REAL
+      -- A4: whole centavos, never a float. See Cents in shared/types/jev.ts.
+      debit             INTEGER,
+      credit            INTEGER
     );
 
     CREATE TABLE IF NOT EXISTS supporting_documents (
